+++ src/extension.ts
@@ -1,19 +1,15 @@
 import * as vscode from 'vscode';
-import * as fs from 'fs';
-import * as path from 'path';
-import { addflowconfig, deleteflowconfig, packproject, pushproject, initproject, addpackageconfig, addlaunchconfig, flowCrendentials } from './flowconfig';
+import { addflowconfig, deleteflowconfig, flowCrendentials, initproject, initprojectforce, packproject, pushproject } from './flowconfig';
 
 export function activate(context: vscode.ExtensionContext) {
-	// const protocols =[{label: 'grpc:'}, {label: 'ws:'}, {label: 'wss:'}, {label: 'pipe:'}, {label: 'socket:'}, {label: 'http:'}, {label: 'https:'}]
-
-	var credentials:flowCrendentials[] = [];
+	var credentials: flowCrendentials[] = [];
 	try {
 		credentials = vscode.workspace.getConfiguration().get<flowCrendentials[]>('openiap.flow.credentials') as flowCrendentials[];
 		// @ts-ignore
-		if(credentials == undefined || credentials.length == 0 || credentials == false) credentials = [];
+		if (credentials == undefined || credentials.length == 0 || credentials == false) credentials = [];
 		var exists = credentials.find(x => x.apidomain == "blahblah");
-		for(var i = 0; i < credentials.length; i++){
-			if(credentials[i].name == "" || credentials[i].name == null){
+		for (var i = 0; i < credentials.length; i++) {
+			if (credentials[i].name == "" || credentials[i].name == null) {
 				credentials[i].name = credentials[i].apidomain;
 			}
 		}
@@ -34,11 +30,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(disposable);
 	disposable = vscode.commands.registerCommand('openiap.initproject', initproject);
 	context.subscriptions.push(disposable);
-	disposable = vscode.commands.registerCommand('openiap.addpackageconfig', addpackageconfig);
-	context.subscriptions.push(disposable);
-	disposable = vscode.commands.registerCommand('openiap.addlaunchconfig', addlaunchconfig);
+	disposable = vscode.commands.registerCommand('openiap.initprojectforce', initprojectforce);
 	context.subscriptions.push(disposable);
 }
-
-// This method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() { }