+++ src/flowconfig.ts
@@ -1,12 +1,9 @@
-import { execSync } from 'child_process';
+import * as fs from 'fs';
+import * as path from 'path';
 import * as vscode from 'vscode';
 import { MultiStepInput } from './multiStepInput';
-import { AppendLineToOutputWindow, GetUser, RequestUserToken, runCommandInOutputWindow, UploadPackage } from './util';
-import * as path from 'path';
-import * as fs from 'fs';
-const ctrossspawn = require('cross-spawn');
+import { AppendLineToOutputWindow, findNPMPath, get, GetUser, RequestUserToken, runCommandInOutputWindow, UploadPackage } from './util';
 const JSON5 = require('json5')
-
 
 const protocols = [{ label: 'grpc:' }, { label: 'ws:' }, { label: 'wss:' }, { label: 'pipe:' }, { label: 'socket:' }, { label: 'http:' }, { label: 'https:' }]
 const yesno = [{ label: 'Yes' }, { label: 'No' }]
@@ -139,7 +136,7 @@ export async function addflowconfig() {
             } else {
                 jwt = await RequestUserToken(prot + '//' + webdomain);
             }
-            
+
             var apiurl = protocol.label + '//' + apidomain + port + pathname;
             if (username != "" && password != "") {
                 apiurl = protocol.label + '//' + username + ':' + password + '@' + apidomain + port + pathname;
@@ -196,7 +193,6 @@ async function InputConfirm(input: MultiStepInput, message: string): Promise<boo
     }
     return false
 }
-
 export async function deleteflowconfig() {
     MultiStepInput.run(async input => {
         try {
@@ -252,7 +248,7 @@ async function pack(): Promise<string> {
         }
         var name = project.name.replace("/", "-").replace("@", "") + "-" + project.version;
         var cmd = findNPMPath();
-        if(cmd != "" && cmd != null) {
+        if (cmd != "" && cmd != null) {
             if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.openiap.json') == true) {
                 var json = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.openiap.json', 'utf8');
                 var tempproject = JSON5.parse(json);
@@ -319,6 +315,167 @@ export async function userSelectConfiguration(): Promise<flowCrendentials | null
         return null;
     }
 }
+var exampleconfig = {
+    "languages": [
+        "nodejs",
+        "browser",
+        "python",
+        "dotnet",
+        "rust",
+        "php",
+        "java"
+    ],
+    "repositories": [
+        {
+            "name": "nodejs",
+            "description": "Node.js example for most SDK commands",
+            "url": "https://github.com/skadefro/nodetest.git"
+        },
+        {
+            "name": "java",
+            "description": "Java example for most SDK commands",
+            "url": "https://github.com/skadefro/javatest.git"
+        },
+        {
+            "name": "python",
+            "description": "Python example for most SDK commands",
+            "url": "https://github.com/skadefro/pythontest.git"
+        },
+        {
+            "name": "dotnet",
+            "description": "C# example for most SDK commands",
+            "url": "https://github.com/skadefro/dotnettest.git"
+        },
+        {
+            "name": "php",
+            "description": "PHP example for most SDK commands",
+            "url": "https://github.com/skadefro/phptest.git"
+        },
+        {
+            "name": "rust",
+            "description": "Rust example with a small CLI for calling some SDK commands",
+            "url": "https://github.com/skadefro/rusttest.git"
+        },
+        {
+            "name": "browser",
+            "description": "Vue3 demo website showing most SDK commands",
+            "url": "https://github.com/openiap/vue3-web-template.git"
+        },
+        {
+            "name": "browser",
+            "description": "AngularJS demo website showing most SDK commands",
+            "url": "https://github.com/open-rpa/openflow-web-angularjs-template.git"
+        },
+        {
+            "name": "browser",
+            "description": "Angular11 demo website showing most SDK commands",
+            "url": "https://github.com/open-rpa/openflow-web-angular11-template.git"
+        },
+        {
+            "name": "browser",
+            "description": "Svelte 4 demo implementation of OpenCores frontend",
+            "url": "https://github.com/openiap/core-web-arch.git"
+        },
+        {
+            "name": "python",
+            "description": "Workitem queue agent example",
+            "url": "https://github.com/openiap/pythontest"
+        }
+    ]
+}
+var exampleconfigupdated = false;
+export async function userSelectLanguage(): Promise<string | null> {
+    try {
+        // Fetch https://raw.githubusercontent.com/openiap/openiap-assistant-repos/refs/heads/main/repositories.json if possible
+        // if not possible, use exampleconfig
+        // if not possible, use hardcoded example
+
+        if (exampleconfigupdated == false) { // only fetch once
+            exampleconfigupdated = true;
+            try {
+                var json = await get("https://raw.githubusercontent.com/openiap/openiap-assistant-repos/refs/heads/main/repositories.json");
+                if (json != null && json != "") {
+                    exampleconfig = JSON.parse(json);
+                }
+            } catch (error) {
+                exampleconfigupdated = true;
+            }
+        }
+        if (exampleconfig.languages.length > 1) {
+            var labels = exampleconfig.languages.map(x => { return { label: x } });
+            const input = vscode.window.createQuickPick();
+            input.title = "Select language";
+            input.step = 1;
+            input.totalSteps = 1;
+            input.placeholder = "Select language to use";
+            input.items = labels
+            input.activeItems = [labels[0]];
+            input.show();
+            await new Promise(resolve => input.onDidAccept(resolve));
+            var selected = input.selectedItems[0];
+            input.hide();
+            var result = exampleconfig.languages.find(x => x == selected.label)
+            return ((result != null && result != undefined) ? result : null)
+        } else {
+            return null;
+        }
+    } catch (error: any) {
+        AppendLineToOutputWindow(error.message);
+        vscode.window.showErrorMessage(error.message);
+        return null;
+    }
+}
+export async function userSelectLanguageProject(): Promise<string | null> {
+    return new Promise<string | null>((resolve, reject) => {
+        MultiStepInput.run(async input => {
+            try {
+                if (exampleconfigupdated == false) { // only fetch once
+                    exampleconfigupdated = true;
+                    try {
+                        var json = await get("https://raw.githubusercontent.com/openiap/openiap-assistant-repos/refs/heads/main/repositories.json");
+                        if (json != null && json != "") {
+                            exampleconfig = JSON.parse(json);
+                        }
+                    } catch (error) {
+                        exampleconfigupdated = true;
+                    }
+                }
+                const languages = exampleconfig.languages.map(x => { return { label: x } });
+                const language = await input.showQuickPick({
+                    title: "Select project language",
+                    step: 1,
+                    totalSteps: 2,
+                    placeholder: 'Choose a language',
+                    items: languages,
+                    activeItem: languages[0],
+                    shouldResume: shouldResume
+                });
+                if (language == null) {
+                    return resolve(null);
+                }
+                const repositories = exampleconfig.repositories.filter(x => x.name == language.label)
+                    .map(x => { return { label: x.description, url: x.url } });
+                const repository = await input.showQuickPick({
+                    title: "Select project",
+                    step: 2,
+                    totalSteps: 2,
+                    placeholder: 'Choose a project',
+                    items: repositories,
+                    activeItem: repositories[0],
+                    shouldResume: shouldResume
+                });
+                if (repository == null) {
+                    return resolve(null);
+                }
+                const result = repositories.find(x => x.label == repository.label);
+                resolve(result?.url ?? null);
+            } catch (error: any) {
+                AppendLineToOutputWindow(error.message);
+                vscode.window.showErrorMessage(error.message);
+            }
+        });
+    });
+}
 export async function pushproject() {
     try {
         if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.json') == false) {
@@ -331,7 +488,7 @@ export async function pushproject() {
         var project = JSON5.parse(json);
         // increase version 
         var version = project.version.split(".");
-        version[version.length -1 ] = (parseInt(version[version.length -1 ]) + 1).toString();
+        version[version.length - 1] = (parseInt(version[version.length - 1]) + 1).toString();
         project.version = version.join(".");
         fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.json', JSON.stringify(project, null, 4));
 
@@ -350,176 +507,41 @@ export async function pushproject() {
         vscode.window.showErrorMessage(error.message);
     }
 }
-export async function addpackageconfig() {
-    var credentials = await userSelectConfiguration();
-    _addpackageconfig(credentials)
-    vscode.window.showInformationMessage(`Project initialised, please edit package.json to your needs`);
-}
-export async function _addpackageconfig(credentials: flowCrendentials | null) {
+export async function _setupProjectRepository(credentials: flowCrendentials | null, forced: boolean) {
     try {
-        if (credentials == null) return;
-
-
-        var pipinstall = false;
-        // is workspace empty ?
         var _files = fs.readdirSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string);
-        if (_files.length == 0) {
-            var pythonpath = findPythonPath();
-            var nodepath = findNodePath();
-            var npmpath = findNodePath();
-            if(pythonpath != "") {
-                pipinstall = true;
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/main.py', `import openiap, asyncio
-from openiap import Client
-async def __wait_for_message(cli:Client, message, payload):
-    workitem = await cli.PopWorkitem("testq")
-    if workitem != None:
-        workitem.state = "successful"
-        cli.UpdateWorkitem(workitem, None, True)
-async def onconnected(cli:Client):
-    try:
-        await cli.Signin()
-        print("Connected to OpenIAP") 
-        queuename = await cli.RegisterQueue("", __wait_for_message)
-        print(f"Consuming queue {queuename}")
-        result = await cli.Query(collectionname="entities", projection={"_created": 1, "name": 1, "_type": 1})
-        print(result)
-        
-    except Exception as e:
-        print(e)
-    # cli.Close()
-async def main():
-    client = openiap.Client()
-    client.onconnected = onconnected
-    await asyncio.sleep(2)
-    while True:
-        await asyncio.sleep(1)
-asyncio.run(main())`);
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/requirements.txt', `openiap`);
-            }
-            if(nodepath != "" && npmpath != "") {
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/main.js', `// @ts-check
-const { openiap } = require("@openiap/nodeapi");
-/** 
- * @param {openiap} client
- * **/
-async function onConnected(client) {
-    var localqueue = await client.RegisterQueue({ queuename:""}, async (msg, payload, user, jwt)=> {
-        var wi = await client.PopWorkitem({"wiq": "testq"});
-        if(wi == null) {
-            console.log("No workitem found");
-            return;
-        }
-        // simulate work
-        await new Promise(resolve => { setTimeout(resolve, 5000) });
-        wi.state = "successful";
-        await client.UpdateWorkitem({workitem: wi});
-        console.log("Updated workitem");
-    })
-    console.log("listening on " + localqueue);
-    var result = await client.Query({query: {}, projection: {"_created":1, "name":1}, top:5})
-    console.log(JSON.stringify(result, null, 2))
-}
-async function main() {
-    var client = new openiap();
-    client.onConnected = onConnected
-    await client.connect();
-}
-main();`);
-            }
-            fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.gitignore', `/node_modules`);
-
-        }
-
-        var json = `{
-            "name": "",
-            "version": "0.0.1",
-            "description": "Example agent, please change",
-            "main": "",
-            "openiap": {
-                "daemon": false,
-                "chromium": false
-            }
-        }`
-        if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.json') == true) {
-            json = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.json', 'utf8');
-        }
-            
-        // get workspace's parent folder name 
-        var parentname = vscode.workspace.workspaceFolders?.[0].uri.fsPath.split("/").pop();
-
-        var project = JSON5.parse(json);
-        if (project.openiap == null) project.openiap = {};
-        if((project.main == null || project.main == "") || (project.openiap.language == null || project.openiap.language == "") ) {
-            var files = await vscode.workspace.findFiles('**/*.ts', '**/node_modules/**');
-            if (files.length == 0) {
-                files = await vscode.workspace.findFiles('**/*.js', '**/node_modules/**');
-                if (files.length > 0) {
-                    project.openiap.language = "nodejs";
-                    project.openiap.typescript = false;
-                }
-            } else {
-                project.openiap.language = "nodejs";
-                project.openiap.typescript = true;
-            }
-            if (files.length == 0) {
-                files = await vscode.workspace.findFiles('**/*.py', '**/node_modules/**');
-                if (files.length > 0) project.openiap.language = "python";
-            } 
-            if (files.length == 0) {
-                files = await vscode.workspace.findFiles('**/*.csproj', '**/node_modules/**');
-                if (files.length > 0) project.openiap.language = "dotnet";
-            }
-            if (files.length == 0) {
-                files = await vscode.workspace.findFiles('**/*.vbproj', '**/node_modules/**');
-                if (files.length > 0) project.openiap.language = "dotnet";
-            }
-            if (files.length == 0) {
-                files = await vscode.workspace.findFiles('**/*.ps1', '**/node_modules/**');
-                if (files.length > 0) project.openiap.language = "powershell";
-            } 
-            if (files.length > 0 && (project.main == null || project.main == "")) {
-                var name:string = files[0].fsPath.replace(vscode.workspace.workspaceFolders?.[0].uri.fsPath as any, "");
-                if(name.startsWith("/") || name.startsWith("\\")) name = name.substring(1);
-                if (project.openiap.language != "dotnet") {
-                    project.main = name;
+        if (_files.length == 0 || forced) {
+            if (credentials == null) return;
+            if (_files.length > 0 && forced) {
+                const input = vscode.window.createQuickPick();
+                input.title = "WARNING! Removed all files ?";
+                input.step = 1;
+                input.totalSteps = 1;
+                input.placeholder = "WARNING, this will delete all existing files in the workspace, are you sure?";
+                input.items = yesno
+                input.activeItems = [yesno[0]];
+                input.show();
+                await new Promise(resolve => input.onDidAccept(resolve));
+                var selected = input.selectedItems[0];
+                input.hide();
+                if (selected.label != 'Yes') {
+                    return;
                 }
             }
-        }
-        var updateapi = false;
-        if (project.name == "" || project.name == null) project.name = parentname;
-        if (project.main?.indexOf(".js") > 0) {
-            if (project.scripts == null) project.scripts = {};
-            if (project.scripts.updateapi == null) {
-                project.scripts.updateapi = "npm uninstall @openiap/nodeapi && npm i @openiap/nodeapi"
-                updateapi = true;
-            } else {
-                if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/node_modules/@openiap/nodeapi') == false) {
-                    updateapi = true;
-                }
-
-            }
-        } else if (project.main?.indexOf(".py") > 0) {
-            if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/requirements.txt') == false) {
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/requirements.txt', `openiap`);
-            } else {
-                var requirements = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/requirements.txt', 'utf8');
-                if (requirements.indexOf('openiap') < 0) {
-                    fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/requirements.txt', requirements + `\nopeniap`);
+            const repo = await userSelectLanguageProject();
+            if (repo == null) return;
+            if (_files.length > 0 && forced) {
+                for (var i = 0; i < _files.length; i++) {
+                    fs.rmSync(path.join(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string, _files[i]), { recursive: true, force: true });
                 }
             }
-        }
-        if (project.openiap.daemon == null) project.openiap.daemon = false;
-        if (project.openiap.chromium == null) project.openiap.chromium = false;
-        fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/package.json', JSON.stringify(project, null, 4));
-
-        if (updateapi) {
-            var cmd = findNPMPath();
-            if(cmd != "" && cmd != null) await runCommandInOutputWindow(cmd, ['run updateapi'], vscode.workspace.workspaceFolders?.[0].uri.fsPath);
-        }
-        if (pipinstall) {
-            var cmd = findPythonPath();
-            if(cmd != "" && cmd != null) await runCommandInOutputWindow(cmd, ['-m pip install openiap'], vscode.workspace.workspaceFolders?.[0].uri.fsPath);
+            const currentfoldername = path.basename(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string);
+            const parent = path.dirname(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string);
+            await runCommandInOutputWindow("git", ["clone", repo, currentfoldername], parent);
+            const gitFolder = path.join(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string, '.git');
+            if (fs.existsSync(gitFolder)) {
+                fs.rmSync(gitFolder, { recursive: true, force: true });
+            }
         }
     } catch (error: any) {
         AppendLineToOutputWindow(error.message);
@@ -529,17 +551,20 @@ main();`);
 export async function initproject() {
     var credentials = await userSelectConfiguration();
     if (credentials == null) return;
-    await _addpackageconfig(credentials)
-    await _addlaunchconfig(credentials)
+    await _setupProjectRepository(credentials, false)
+    await _ensureenv(credentials)
     vscode.window.showInformationMessage(`Project initialised, please edit package.json to your needs`);
 
 }
-export async function addlaunchconfig() {
+export async function initprojectforce() {
     var credentials = await userSelectConfiguration();
-    _addlaunchconfig(credentials)
-    vscode.window.showInformationMessage(`Launch checked`);
+    if (credentials == null) return;
+    await _setupProjectRepository(credentials, true)
+    await _ensureenv(credentials)
+    vscode.window.showInformationMessage(`Project initialised, please edit package.json to your needs`);
+
 }
-export async function _addlaunchconfig(credentials: flowCrendentials | null) {
+export async function _ensureenv(credentials: flowCrendentials | null) {
     try {
         if (credentials == null) return;
         var project: any = {}
@@ -550,224 +575,75 @@ export async function _addlaunchconfig(credentials: flowCrendentials | null) {
             vscode.window.showInformationMessage(`package.json missing, please run initproject first`);
             return;
         }
-        // make sure .vscode filder exists and check if launch.json exists
-        if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode') == false) {
-            fs.mkdirSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode');
-        }
-        if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json') == false) {
-            var _files = fs.readdirSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath as string);
-            if(_files.indexOf('main.py') > 0 && _files.indexOf('main.js') > 0) {
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', `{
-                    "version": "0.2.0",
-                    "configurations": [
-                        {
-                            "type": "node",
-                            "request": "launch",
-                            "name": "NodeJS: Run main.js",
-                            "skipFiles": [
-                                "<node_internals>/**"
-                            ],
-                            "program": "\${workspaceFolder}/main.js",
-                            "console": "integratedTerminal",
-                            "env": {
-                                "apiurl": "${credentials.apiurl}",
-                                "jwt": "${credentials.jwt}"
-                        
-                            }
-                        },
-                        {
-                            "name": "Python: Run main.py",
-                            "type": "python",
-                            "request": "launch",
-                            "program": "main.py",
-                            "console": "integratedTerminal",
-                            "env": {
-                                "apiurl": "${credentials.apiurl}",
-                                "jwt": "${credentials.jwt}"
-                        
-                            }
-                        }
-                    ]
-                }`);
-            } else {
-                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', `{
-                    "version": "0.2.0",
-                    "configurations": [
-                    ]
-                }`);
-            }
-            fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/.gitignore', `launch.json`);
-
-        }
-        var configuration: any = null;
-        var json = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', 'utf8');
-        var launch = JSON5.parse(json);
-        for(var i = 0; i < launch.configurations.length; i++) {
-            try {
-                if(launch.configurations[i].program == null) continue;
-                if(launch.configurations[i].program.toLowerCase().indexOf(project.main?.toLowerCase()) > 0) {
-                    configuration = launch.configurations[i];
-                } else if(launch.configurations[i].program.toLowerCase().indexOf(`${project.name?.toLowerCase()}.dll`) > 0) {
-                    configuration = launch.configurations[i];
-                } else if(launch.configurations[i].program.toLowerCase().indexOf(`.dll`) > 0) {
-                    configuration = launch.configurations[i];
-                }
-            } catch (error) {
-                console.log(error);
-            }
-        }
-        if(configuration == null && launch.configurations.length == 1) configuration = launch.configurations[0];
-        if(configuration == null) {
-            if (project.openiap.language == "python") {
-                configuration = {
-                    "name": "Python: Run "+ project.main,
-                    "type": "python",
-                    "request": "launch",
-                    "program": project.main,
-                    "console": "integratedTerminal",
-                };
-                launch.configurations.push(configuration);
-            } else if (project.openiap.language == "nodejs") {
-                if(project.openiap.typescript == true && project.main.endsWith(".ts")) {
-                    configuration =         {
-                        "type": "node",
-                        "request": "launch",
-                        "name": "TypeScript: Run " + project.main,
-                        "preLaunchTask": "tsc: build - tsconfig.json",
-                        "runtimeArgs": [
-                            "--loader",
-                            "ts-node/esm"
-                        ],
-                        "skipFiles": [
-                            "<node_internals>/**"
-                        ],
-                        "program": `\${workspaceFolder}/${project.main}`,
-                        "console": "integratedTerminal"
+        let usedlaunch = false;
+        if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode') == true) {
+            if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json') == true) {
+                var json = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', 'utf8');
+                var launch = JSON5.parse(json);
+                for (var i = 0; i < launch.configurations.length; i++) {
+                    var conf = launch.configurations[i];
+                    if (conf.program == null) continue;
+                    if (conf != null && conf.env != null && conf.env?.apiurl != null && conf.env?.jwt != null) {
+                        conf.env.apiurl = credentials.apiurl;
+                        conf.env.jwt = credentials.jwt;
+                        usedlaunch = true;
                     }
-                    // not needed anymore
-                    // var files = await vscode.workspace.findFiles('**/nodemon.json', '**/node_modules/**');
-                    // if(files.length == 0) {
-                    //     fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/nodemon.json', `{
-                    //         "execMap": {
-                    //             "ts": "ts-node"
-                    //         }
-                    //     }`);
-                    // }
-                } else {
-                    configuration = {
-                        "type": "node",
-                        "request": "launch",
-                        "name": "NodeJS: Run " + project.main,
-                        "skipFiles": [
-                            "<node_internals>/**"
-                        ],
-                        "program": `\${workspaceFolder}/${project.main}`,
-                        "console": "integratedTerminal",
-                    };
                 }
-                launch.configurations.push(configuration);
-            } else if (project.openiap.language == "dotnet") {
-                configuration = {
-                    "name": ".NET Core Launch ${project.name}",
-                    "type": "coreclr",
-                    "request": "launch",
-                    "preLaunchTask": "build",
-                    "program": `\${workspaceFolder}/bin/Debug/net6.0/${project.name}.dll`,
-                    "console": "integratedTerminal",
-                    "stopAtEntry": false,
-                    "cwd": "${workspaceFolder}",
-                };
-                launch.configurations.push(configuration);
-            } else if (project.openiap.language == "powershell") {
-                configuration =        {
-                    "name": "PowerShell: Run ${project.main}",
-                    "type": "PowerShell",
-                    "request": "launch",
-                    "script": project.main,
-                    "args": [],
-                    "cwd": "${file}"
-                };
-                launch.configurations.push(configuration);
+                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', JSON.stringify(launch, null, 4));
             }
         }
-        if(configuration == null) {
-            vscode.window.showInformationMessage(`Error adding launch.json`);
-            return;
-        }
-        if(configuration.env == null) configuration.env = {};
-        configuration.env.apiurl = credentials.apiurl;
-        configuration.env.jwt = credentials.jwt;
-        for(var i = 0; i < launch.configurations.length; i++) {
-            var conf = launch.configurations[i];
-            if(conf.program == null) continue;
-            if(conf != null&& conf.env != null && conf.env?.apiurl != null && conf.env?.jwt != null) {
-                conf.env.apiurl = credentials.apiurl;
-                conf.env.jwt = credentials.jwt;
+        if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.env') == true || usedlaunch == false) {
+            var lines: string[] = [];
+            let env = "";
+            if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.env') == true) {
+                env = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.env', 'utf8');
+                lines = env.split("\n");
+            }
+            for (var i = 0; i < lines.length; i++) {
+                var line = lines[i];
+                if (line.trim().startsWith("apiurl=")) {
+                    var parts = line.split("=");
+                    if (parts.length > 1) {
+                        line = "apiurl=" + credentials.apiurl;
+                        lines[i] = line;
+                    }
+                }
+                if (line.trim().startsWith("jwt=")) {
+                    var parts = line.split("=");
+                    if (parts.length > 1) {
+                        line = "jwt=" + credentials.jwt;
+                        lines[i] = line;
+                    }
+                }
+            }
+            if (env.indexOf("apiurl=") == -1) {
+                lines.push("apiurl=" + credentials.apiurl);
+            }
+            if (env.indexOf("jwt=") == -1) {
+                lines.push("jwt=" + credentials.jwt);
+            }
+            fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.env', lines.join("\n"));
+            if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.gitignore') == false) {
+                fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.gitignore', '.env');
+            } else {
+                var gitignore = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.gitignore', 'utf8');
+                if (gitignore.indexOf('.env') == -1) {
+                    gitignore += '\n.env';
+                    fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.gitignore', gitignore);
+                }
+            }
+            if (fs.existsSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.npmignore') == true) {
+                var npmignore = fs.readFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.npmignore', 'utf8');
+                if (npmignore.indexOf('.env') == -1) {
+                    npmignore += '\n.env';
+                    fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.npmignore', npmignore);
+                }
             }
         }
-        fs.writeFileSync(vscode.workspace.workspaceFolders?.[0].uri.fsPath + '/.vscode/launch.json', JSON.stringify(launch, null, 4));
+
+
     } catch (error: any) {
         AppendLineToOutputWindow(error.message);
         vscode.window.showErrorMessage(`Error adding launch.json ` + error.message);
-    }
-}
-
-export function findPythonPath() {
-    var result = findInPath("python3")
-    if (result == "") result = findInPath("python")
-    return result;
-}
-export function findDotnetPath() {
-    return findInPath("dotnet")
-}
-export function findXvfbPath() {
-    return findInPath("xvfb-run")
-}
-export function findNodePath() {
-    return findInPath("node")
-}
-export function findNPMPath() {
-    const child = (process.platform === 'win32' ? 'npm.cmd' : 'npm')
-    return findInPath(child)
-}
-export function findChromiumPath() {
-    var result = findInPath("chromium-browser");
-    if (result == "") result = findInPath("chromium");
-    return result
-}
-export function findChromePath() {
-    var result = findInPath("google-chrome");
-    if (result == "") result = findInPath("chrome");
-    return result
-}
-export function findInPath(exec:string):string | null {
-    try {
-        let command;
-        switch (process.platform) {
-            case 'linux':
-            case 'darwin':
-                command = 'which';
-                break;
-            case 'win32':
-                command = 'where.exe';
-                break;
-            default:
-                throw new Error(`Unsupported platform: ${process.platform}`);
-        }
-        const result:any = ctrossspawn.sync(command, [exec], { stdio: 'pipe' });
-        if (result.status === 0) {
-            const stdout = result.stdout.toString();
-            const lines = stdout.split(/\r?\n/).filter((line:string) => line.trim() !== '')
-                .filter((line:string) => line.toLowerCase().indexOf("windowsapps\\python3.exe") == -1)
-                .filter((line:string) => line.toLowerCase().indexOf("windowsapps\\python.exe") == -1);
-            if(lines.length > 0)  return lines[0]
-        } else {
-            if(result.stderr != null) console.log(result.stderr.toString());
-            if(result.stdout != null) console.log(result.stdout.toString());
-        }
-        return "";
-    } catch (error) {
-        return "";
-        // throw error;
     }
 }